+++ clients/admin-ui/src/features/data-discovery-and-detection/action-center/tables/DiscoveredSystemAggregateTable.tsx
@@ -39,9 +39,14 @@ import {
 } from "~/features/data-discovery-and-detection/action-center/action-center.slice";
 import useActionCenterTabs, {
   getIndexFromHash,
-} from "~/features/data-discovery-and-detection/action-center/tables/useActionCenterTabs";
-import { successToastContent } from "~/features/data-discovery-and-detection/action-center/utils/successToastContent";
-import { DiffStatus, SystemStagedResourcesAggregateRecord } from "~/types/api";
+} from "~/features/data-discovery-and-detection/action-center/hooks/useActionCenterTabs";
+import { SuccessToastContent } from "~/features/data-discovery-and-detection/action-center/SuccessToastContent";
+import {
+  ConsentStatus,
+  ConsentStatusInfo,
+  DiffStatus,
+  SystemStagedResourcesAggregateRecord,
+} from "~/types/api";
 import { isErrorResult } from "~/types/errors";
 
 import { DebouncedSearchInput } from "../../../common/DebouncedSearchInput";
@@ -56,6 +61,10 @@ export const DiscoveredSystemAggregateTable = ({
 }: DiscoveredSystemAggregateTableProps) => {
   const router = useRouter();
   const tabHash = router.asPath.split("#")[1];
+
+  const [firstItemConsentStatus, setFirstItemConsentStatus] = useState<
+    ConsentStatusInfo | null | undefined
+  >();
 
   const {
     PAGE_SIZES,
@@ -110,6 +119,17 @@ export const DiscoveredSystemAggregateTable = ({
     }
   }, [data, setTotalPages]);
 
+  useEffect(() => {
+    if (data?.items && !firstItemConsentStatus) {
+      // this ensures that the column header remembers the consent status
+      // even when the user navigates to a different paginated page
+      const consentStatus = data.items.find(
+        (item) => item.consent_status?.status === ConsentStatus.ALERT,
+      )?.consent_status;
+      setFirstItemConsentStatus(consentStatus);
+    }
+  }, [data, firstItemConsentStatus]);
+
   const handleTabChange = (index: number) => {
     onTabChange(index);
     setRowSelection({});
@@ -120,6 +140,7 @@ export const DiscoveredSystemAggregateTable = ({
     onTabChange: handleTabChange,
     readonly: actionsDisabled,
     allowIgnore: !activeParams.diff_status.includes(DiffStatus.MUTED),
+    consentStatus: firstItemConsentStatus,
   });
 
   const tableInstance = useReactTable({
@@ -167,7 +188,7 @@ export const DiscoveredSystemAggregateTable = ({
     } else {
       toast(
         successToastParams(
-          successToastContent(
+          SuccessToastContent(
             `${totalUpdates} assets have been added to the system inventory.`,
             () => router.push(SYSTEM_ROUTE),
           ),
@@ -195,7 +216,7 @@ export const DiscoveredSystemAggregateTable = ({
     } else {
       toast(
         successToastParams(
-          successToastContent(
+          SuccessToastContent(
             `${totalUpdates} assets have been ignored and will not appear in future scans.`,
             () => onTabChange(getIndexFromHash("#ignored")!),
           ),