+++ clients/admin-ui/src/types/api/models/AggregatedConsent.ts
@@ -1,0 +1,10 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+export enum AggregatedConsent {
+  WITH_CONSENT = "with_consent",
+  WITHOUT_CONSENT = "without_consent",
+  EXEMPT = "exempt",
+  UNKNOWN = "unknown",
+}